requirements.txt
streamlit
langchain
faiss-cpu
PyPDF2
python-docx
pandas
pytesseract
Pillow
whisper
openai
pdfplumber
pytube==15.0.0        
moviepy==1.0.3        
whisper==0.1.4         
torch==2.0.1          
opencv-python==4.8.0.74
pydub==0.25.1        
ffmpeg-python==0.2.0  
numpy==1.24.4  
